requirements.txt
aniso8601==1.2.0
appnope==0.1.0
backports-abc==0.4
beautifulsoup4==4.5.3
click==6.6
cloudpickle==0.2.1
cycler==0.10.0
cymem==1.31.2
Cython==0.25.2
dask==0.13.0
decorator==4.0.10
Flask==0.11.1
Flask-RESTful==0.3.5
futures==3.0.5
gnupg==2.0.2
h5py==2.6.0
ipykernel==4.0.3
ipython==5.2.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
isodate==0.5.4
itsdangerous==0.24
Jinja2==2.8
joblib==0.10.3
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
keepalive==0.5
Keras==1.2.0
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
Morfessor==2.0.2a3
murmurhash==0.26.4
nbconvert==4.2.0
nbformat==4.1.0
ner==0.1
networkx==1.11
nltk==3.2.1
notebook==4.3.1
numpy==1.12.0
pandas==0.19.0
path.py==0.0.0
pathlib==1.0.1
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.4.2
plac==0.9.6
polyglot==16.7.4
pomegranate==0.4.0
preshed==0.46.4
prompt-toolkit==1.0.8
protobuf==3.1.0.post1
psutil==4.4.2
ptyprocess==0.5
py==1.4.31
py-entitymatching==0.1.0
py-stringmatching==0.2.1
py-stringsimjoin==0.1.0
pycld2==0.31
Pygments==2.1.3
PyICU==1.9.5
pyparsing==2.1.4
PyPrind==2.9.8
pyspotlight==0.7.1
pytest==3.0.5
python-dateutil==2.5.3
python-firebase==1.2
python-Levenshtein==0.12.0
pytz==2016.7
PyYAML==3.12
pyzmq==14.7.0
qtconsole==4.2.1
rdflib==4.2.1
requests==2.2.1
scikit-image==0.12.3
scikit-learn==0.18
scipy==0.18.1
seaborn==0.7.1
selenium==3.0.2
semver==2.7.2
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
spacy==1.3.0
SPARQLWrapper==1.7.6
sputnik==0.9.3
tensorflow==0.12.1
terminado==0.5
Theano==0.8.2
thinc==5.0.8
toolz==0.8.2
tornado==4.4.2
traitlets==4.3.1
ujson==1.35
uuid==1.30
wcwidth==0.1.7
Werkzeug==0.11.11
widgetsnbextension==1.2.6
yandex.translate==0.3.5
